+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- **BC Break!**: Updated `PdfDocument` class to take account of the new
+  colors model.
+- Added `PdfColorInterface`, `PdfRgbColor` and `PdfGrayColor`.
 - Updated Symfony to version 7.2.0.
 - Updated Symfony to version 7.1.9.